add conditional rebuild, with a check every 30s

+++ run.sh
@@ -14,41 +14,66 @@ if [ -z "$path" ]; then
   path="/"
 fi
 
+git config --global advice.detachedHead false
+
+last_commit=""
+
+function jekyll_build {
+  echo "Building web content..."
+  cd "/target/$path"
+  # determine whether we should use jekyll
+  if [ -f ".nojekyll" ]; then
+    echo "Found .nojekyll file in root of repo - not using jekyll to process the site."
+  else
+    echo "Using jekyll to build the site..."
+
+    bundle config set --local system 'true'
+    bundle install
+
+    bundle exec jekyll build --incremental
+    #bundle exec jekyll serve --host 0.0.0.0 --incremental --watch
+
+    if [ -d "_site" ]; then
+      # move to the jekyll generated site (for the next copy command)
+      cd "_site"
+    else
+      echo "WARNING: Failed to generate site using Jekyll - serving the publication path anyway as a fallback" >&2
+    fi
+  fi
+  # copy all files from the site to the web root
+  cp -rv "/target"/* "$webroot"
+}
+
+function update_repo {
+  cd "/target"
+  # check out the repo and the specific publication source
+  git fetch -p
+  latest_commit=$(git log --all --oneline | head -1)
+  if [ "$latest_commit" != "$last_commit" ]; then
+    echo "New commits have been made since we last checked - updating repo"
+    git checkout "origin/$branch"
+    git reset --hard HEAD
+    git clean -d --force
+    git submodule update --init
+
+    last_commit=$(git log --all --oneline | head -1)
+
+    jekyll_build
+  else
+    echo "No changes since our last check - no need for rebuild"
+  fi
+}
+echo
 echo "Repository: $repo_url, branch: $branch, path: $path"
 
 # check out the repo and the specific publication source
-git config --global advice.detachedHead false
 git clone "$repo_url" "/target"
-cd "/target"
-git fetch -p
-git checkout "origin/$branch"
-git reset --hard HEAD
-git clean -d --force
-git submodule update --init
-
-cd ./"$path"
-# determine whether we should use jekyll
-if [ -f ".nojekyll" ]; then
-  echo "Found .nojekyll file in root of repo - not using jekyll to process the site."
-else
-  echo "Using jekyll to build the site..."
-
-  bundle config set --local system 'true'
-  bundle install
-
-  bundle exec jekyll build --incremental
-  #bundle exec jekyll serve --host 0.0.0.0 --incremental --watch
-
-  if [ -d "_site" ]; then
-    # move to the jekyll generated site (for the next copy command)
-    cd "_site"
-  else
-    echo "WARNING: Failed to generate site using Jekyll - serving the publication path anyway as a fallback" >&2
-  fi
+if [ "$?" != 0 ]; then
+  echo "Unable to clone repository $repo_url - cannot continue" >&2
+  exit 1
 fi
-# copy all files from the site to the web root
 mkdir -vp "$webroot"
-cp -rv ./* "$webroot"
+update_repo
 
 # launch the webserver in the background - see nginx_site.conf for site configuration
 nginx
@@ -64,5 +89,5 @@ cd "/target"
 # monitor for any upstream changes and rebuild
 while true; do
   sleep 30
-  # TODO
+  update_repo
 done
